pyproject.toml
[project]
name = "promql-agent"
version = "0.1.0"
description = "CLI agent for Prometheus, Loki, and Grafana"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph-sdk",
    "llama-api-client",
    "python-dotenv",
    "requests",
    "prometheus-api-client",
    "httpx",
]

[project.scripts]
promql-agent = "cli:main"
cli = "cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend/app"]

[tool.hatch.build.targets.wheel.sources]
"backend/app" = "app"

[tool.hatch.build.targets.wheel.force-include]
"cli.py" = "cli.py"

[tool.uv]
dev-dependencies = [
    "pytest",
    "pytest-asyncio",
]
